Move jobserver integration into separate crate
Future commits on this branch will make our implementation pretty
specific to rustc (i.e., talking to Cargo and such) so coupling
ourselves into rustc_data_structures which is intended to be more
"agnostic" isn't a great fit.

This also removes uses of the jobserver inside rustc that refer directly
to the `jobsever` crate (and similar cleanups will follow in the next
few commits).