Track whether to emit jobserver messages
This adds the -Zjobserver-token-requests debugging flag which Cargo will
pass down to us, but for now that flag doesn't do anything.

We don't really have a good way to deal with multiple initialize calls
so for now this implementation just ignores all but the first
construction's token requesting mode. It's plausible we should assert
that all calls have the same mode, but it's unclear whether that really
brings us any benefits (and it does have a cost of forcing people to
thread things through on their end).